move to session component
