+++ wanikani/static/js/components/session/constants.js
