+++ wanikani/static/js/components/session/helpers.js
